wip: refactor and allow object tables to reset row values
